# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "忆述光华 API",
    "description": "个人记忆整理与回顾后端。功能：认证（微信/手机号/JWT）、内容入库（照片/文字/语音）、四层事件聚合、描述性搜索（RAG）、同步（LWW）、AI 分类/画像/护栏。契约对齐《忆述光华_Sprint1规划.md》S1-02。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/events/timeline": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Timeline",
        "description": "时间轴（F8）：L1 日卡片 + L2 主题事件 + L3 主题流；按 start_time 排序\n\n2026-08-20：事件聚合服务已接线（services/events.py），返回真实 events 数据；\n手动操作（merge/split/confirm）仍 501（用户操作优先原则，后续实现）。\n安全：需登录。",
        "operationId": "timeline_api_v1_events_timeline_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_EventOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/merge": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Merge Events",
        "description": "用户合并（B3-5：存合并规则，算法永不覆盖用户决定，AGG-013）\n\n2026-08-20：手动操作已接线；source 内容并入 target，source 软删，target 置 confirmed。",
        "operationId": "merge_events_api_v1_events_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_EventOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/events/split": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Split Event",
        "description": "用户拆分（B3-5）：拆出内容建独立事件；拆出的新事件置 confirmed",
        "operationId": "split_event_api_v1_events_split_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventSplitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_EventOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/events/confirm": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Confirm Event",
        "description": "用户确认（置信度<0.7 转正；用户背书后不再被算法改动）",
        "operationId": "confirm_event_api_v1_events_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_EventOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/interview/answers": {
      "post": {
        "tags": [
          "interview"
        ],
        "summary": "Interview Answers",
        "operationId": "interview_answers_api_v1_interview_answers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewAnswers"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_InterviewResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/interview/profile": {
      "get": {
        "tags": [
          "interview"
        ],
        "summary": "Interview Profile",
        "operationId": "interview_profile_api_v1_interview_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "List Messages",
        "description": "消息列表（sent_at DESC 游标分页；status 过滤 unread/read/archived）",
        "operationId": "list_messages_api_v1_messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Page_MessageOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/{msg_id}/read": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Mark Read",
        "description": "单条已读（越权访问他人消息 → 404 不泄露存在性）",
        "operationId": "mark_read_api_v1_messages__msg_id__read_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "msg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Msg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/read-all": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Mark All Read",
        "description": "全部已读（消息中心列表页退出即读，不打断）",
        "operationId": "mark_all_read_api_v1_messages_read_all_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse",
        "description": "统一响应体：{code, message, data, request_id}"
      },
...
      "ApiResponse_EventOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[EventOut]"
      },
...
      "ApiResponse_InterviewResult_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[InterviewResult]"
      },
...
      "ApiResponse_Page_MessageOut__": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page_MessageOut_"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[Page[MessageOut]]"
      },
...
      "ApiResponse_list_EventOut__": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EventOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[list[EventOut]]"
      },
...
      "EventConfirmRequest": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "event_id"
        ],
        "title": "EventConfirmRequest",
        "description": "用户确认事件（置信度<0.7 转正；用户背书后算法不再改动）"
      },
      "EventMergeRequest": {
        "properties": {
          "target_event_id": {
            "type": "string",
            "title": "Target Event Id"
          },
          "source_event_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Source Event Ids"
          }
        },
        "type": "object",
        "required": [
          "target_event_id",
          "source_event_ids"
        ],
        "title": "EventMergeRequest",
        "description": "用户手动合并事件（B3-5：存合并规则，用户操作优先）"
      },
      "EventOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "level": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Level"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "title_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Source"
          },
          "cover_content_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Content Id"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "emotion": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotion"
          },
          "sensitivity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensitivity"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "generated_by": {
            "type": "string",
            "title": "Generated By"
          },
          "content_count": {
            "type": "integer",
            "title": "Content Count",
            "default": 0
          },
          "photo_count": {
            "type": "integer",
            "title": "Photo Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "level",
          "title",
          "title_source",
          "cover_content_id",
          "start_time",
          "end_time",
          "place",
          "emotion",
          "sensitivity",
          "confidence",
          "status",
          "generated_by"
        ],
        "title": "EventOut"
      },
      "EventSplitRequest": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "content_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Content Ids"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "content_ids"
        ],
        "title": "EventSplitRequest",
        "description": "用户手动拆分事件"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterviewAnswers": {
        "properties": {
          "answers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Answers",
            "description": "三问答案：{\"important_person\": \"...\", \"life_turn\": \"...\", \"proud_thing\": \"...\"}"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "InterviewAnswers"
      },
      "InterviewResult": {
        "properties": {
          "dimensions": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Dimensions"
          },
          "confirmation": {
            "type": "string",
            "title": "Confirmation"
          }
        },
        "type": "object",
        "required": [
          "dimensions",
          "confirmation"
        ],
        "title": "InterviewResult"
      },
      "MessageOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "msg_type": {
            "type": "string",
            "title": "Msg Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "sent_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sent At"
          },
          "read_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "channel",
          "msg_type",
          "title",
          "body",
          "status",
          "sent_at"
        ],
        "title": "MessageOut",
        "description": "消息出参（in-app 列表项 / push 记录）"
      },
...
      "Page_MessageOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "type": "array",
            "title": "Items"
          },
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[MessageOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}